# lib/api-zod/src/generated/types/referralRedeemInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReferralRedeemInput {
  /**
     * @minLength 4
     * @maxLength 16
     */
  code: string;
}