# lib/api-zod/src/generated/types/referralSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReferralSummaryRedemption } from './referralSummaryRedemption';
import type { ReferralSummaryStats } from './referralSummaryStats';

export interface ReferralSummary {
  /**
     * @minLength 4
     * @maxLength 16
     */
  code: string;
  inviteUrl: string;
  rewardDays: number;
  stats: ReferralSummaryStats;
  /** The caller's own redemption of someone else's code, if any. */
  redemption: ReferralSummaryRedemption;
}